make reset_pyvista and dynamic_scraper serializable by outsourcing them
